{,un}packNetworkAddress improvements, relocation.
* Move {,un}packNetworkAddress to netutil.py
* Make {,un}packNetworkAddress check existence
  of {inet_pton,inet_ntop} instead of platform
